requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.0
isort==5.13.2
mypy==1.10.1
pre-commit==3.7.1
python-decouple==3.8
